Baked the Decidability of the Adj Relation into RamseyGraphProp
The fact that we are dealing now with graphs and not colorings of a
complete graph means that we can have wild adjacency relations, in
particular adjacency relations that are not decidable (as opposed to
one fixed adjacency relation in the coloring case) so in order to keep
things "normal" let us just require decidability from the definition.

Fixed all the things this change broke.