Finished the Proof of R 3 3 3 = 17
Note that this requires an even higher upper limit of heartbeats to
make the native_decide trick work. The proof can be significantly
sped-up by considering only the lower-diagonal cases and using the
interval_cases tactic.

Another improvement would be putting the sSup condition in
RamseyPropG6Partition, but this works for now.

Closes #3