Change Ramsey Definition from sInf to Nat.find
Everything in finite Ramsey theory is decidable/computable so it is
unfortunate to define Ramsey numbers as noncomputable. In Nat, sInf is
actually defined as Nat.find, which is computable under some
conditions (that we meet) so we go with that.

The changes are more or less one-to-one: sInf_mem -> Nat.find_spec,
sInf_le -> Nat.find_le. The only thing missing was a theorem for
upward-closed properties which we added to Utils but should probably
exist in Mathlib.

Also, to use Nat.find we had to reorder some of the lemmas and prove a
few instances of decidability.